Handle non-files in the the `ReadRead` implementation for `std::fs::File`.
`File` can open special devices like /dev/random or /dev/null; make sure
those are handled properly in `ReadReady`.

Also, correctly report the number of ready bytes if the current position
is in the middle of a file.
