+++ src/fs/mod.rs
@@ -3,8 +3,6 @@
 mod fd_flags;
 mod file_io_ext;
 
-use crate::io::IsReadWrite;
-
 pub use fd_flags::{FdFlags, GetSetFdFlags, SetFdFlags};
 pub use file_io_ext::{Advice, FileIoExt};
 
@@ -16,20 +14,3 @@ pub use file_io_ext::{Advice, FileIoExt};
 // TODO: test that remove_dir can remove symlink_dirs, and remove_file files.
 
 // TODO: poll
-
-impl crate::io::ReadReady for std::fs::File {
-    #[inline]
-    fn num_ready_bytes(&self) -> std::io::Result<u64> {
-        let (read, _write) = self.is_read_write()?;
-        if read {
-            let metadata = self.metadata()?;
-            if metadata.is_file() {
-                return Ok(metadata.len());
-            }
-        }
-        Err(std::io::Error::new(
-            std::io::ErrorKind::InvalidInput,
-            "stream is not readable",
-        ))
-    }
-}
